+++ packages/javascript-kernel-extension/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyterlite/kernel",
+  "name": "@jupyterlite/javascript-kernel-extension",
   "version": "0.1.0",
-  "description": "JupyterLite - Kernel",
+  "description": "JupyterLite - JavaScript Kernel Extension",
   "homepage": "https://github.com/jtpio/jupyterlite",
   "bugs": {
     "url": "https://github.com/jtpio/jupyterlite/issues"
@@ -32,27 +32,17 @@
   ],
   "scripts": {
     "build": "tsc -b",
-    "build:test": "tsc --build tsconfig.test.json",
     "clean": "rimraf lib && rimraf tsconfig.tsbuildinfo",
     "docs": "typedoc src",
     "prepublishOnly": "npm run build",
-    "test": "jest",
-    "test:cov": "jest --collect-coverage",
-    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
-    "test:debug:watch": "node --inspect-brk node_modules/.bin/jest --runInBand --watch",
     "watch": "tsc -b --watch"
   },
   "dependencies": {
-    "@jupyterlab/services": "^6.0.5",
-    "@jupyterlite/kernelspec": "^0.1.0",
-    "@lumino/coreutils": "^1.5.3",
-    "@lumino/disposable": "^1.4.3",
-    "@lumino/signaling": "^1.4.3"
+    "@jupyterlite/kernelspec": "^0.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.12.1",
-    "@jupyterlab/testutils": "^3.0.0",
     "@types/jest": "^26.0.10",
     "jest": "^26.4.2",
     "rimraf": "~3.0.0",
