+++ packages/kernel/src/index.ts
@@ -3,9 +3,4 @@
 
 export * from './kernels';
 export * from './default';
-export * from './registry';
 export * from './tokens';
-
-import * as Kernel from './kernel';
-
-export { Kernel };
