+++ packages/kernel/src/tokens.ts
@@ -1,22 +1,15 @@
-import { Kernel } from '@jupyterlab/services';
+import { Kernel, KernelMessage } from '@jupyterlab/services';
 
 import { Token } from '@lumino/coreutils';
 
-import { Kernels } from './kernels';
+import { IObservableDisposable } from '@lumino/disposable';
 
-import { KernelRegistry } from './registry';
+import { Kernels } from './kernels';
 
 /**
  * The token for the kernels service.
  */
 export const IKernels = new Token<IKernels>('@jupyterlite/kernel:IKernels');
-
-/**
- * The token for the kernel registry.
- */
-export const IKernelRegistry = new Token<IKernelRegistry>(
-  '@jupyterlite/kernel:IKernelRegistry'
-);
 
 /**
  * An interface for the Kernels service.
@@ -31,11 +24,126 @@ export interface IKernels {
 }
 
 /**
- * An interface for the KernelRegistry service.
+ * An interface for a kernel running in the browser.
  */
-export interface IKernelRegistry {
+export interface IKernel extends IObservableDisposable {
   /**
-   * Register a new kernel.
+   * The id of the server-side kernel.
    */
-  register: (options: KernelRegistry.IKernelOptions) => void;
+  readonly id: string;
+
+  /**
+   * Handle an incoming message from the client.
+   *
+   * @param msg The message to handle
+   */
+  handleMessage(msg: KernelMessage.IMessage): Promise<void>;
+
+  /**
+   * Handle a `kernel_info_request` message.
+   *
+   * @returns A promise that resolves with the kernel info.
+   */
+  kernelInfoRequest(): Promise<KernelMessage.IInfoReplyMsg['content']>;
+
+  /**
+   * Handle an `execute_request` message.
+   *
+   * @param content - The content of the execute_request kernel message
+   */
+  executeRequest(
+    content: KernelMessage.IExecuteRequestMsg['content']
+  ): Promise<KernelMessage.IExecuteResultMsg['content']>;
+
+  /**
+   * Handle a `complete_request` message.
+   *
+   * @param content - The content of the request.
+   */
+  completeRequest(
+    content: KernelMessage.ICompleteRequestMsg['content']
+  ): Promise<KernelMessage.ICompleteReplyMsg['content']>;
+
+  /**
+   * Handle an `inspect_request` message.
+   *
+   * @param content - The content of the request.
+   *
+   * @returns A promise that resolves with the response message.
+   */
+  inspectRequest(
+    content: KernelMessage.IInspectRequestMsg['content']
+  ): Promise<KernelMessage.IInspectReplyMsg['content']>;
+
+  /**
+   * Send a `history_request` message.
+   *
+   * @param content - The content of the request.
+   *
+   * @returns A promise that resolves with the response message.
+   */
+  historyRequest(
+    content: KernelMessage.IHistoryRequestMsg['content']
+  ): Promise<KernelMessage.IHistoryReplyMsg['content']>;
+
+  /**
+   * Handle an `is_complete_request` message.
+   *
+   * @param content - The content of the request.
+   *
+   * @returns A promise that resolves with the response message.
+   */
+  isCompleteRequest(
+    content: KernelMessage.IIsCompleteRequestMsg['content']
+  ): Promise<KernelMessage.IIsCompleteReplyMsg['content']>;
+
+  /**
+   * Handle a `comm_info_request` message.
+   *
+   * @param content - The content of the request.
+   *
+   * @returns A promise that resolves with the response message.
+   */
+  commInfoRequest(
+    content: KernelMessage.ICommInfoRequestMsg['content']
+  ): Promise<KernelMessage.ICommInfoReplyMsg['content']>;
+
+  /**
+   * Send an `input_request` message.
+   *
+   * @param content - The content of the request.
+   */
+  inputRequest(
+    content: KernelMessage.IInputRequestMsg['content']
+  ): Promise<void>;
+}
+
+/**
+ * A namespace for IKernel statics.
+ */
+export namespace IKernel {
+  /**
+   * The type for the send message function.
+   */
+  export type SendMessage = (msg: KernelMessage.IMessage) => void;
+
+  /**
+   * The instantiation options for an IKernel.
+   */
+  export interface IOptions {
+    /**
+     * The kernel id.
+     */
+    id: string;
+
+    /**
+     * The session id.
+     */
+    sessionId: string;
+
+    /**
+     * The method to send messages back to the server.
+     */
+    sendMessage: SendMessage;
+  }
 }
