+++ packages/kernelspec/package.json
@@ -43,6 +43,7 @@
     "watch": "tsc -b --watch"
   },
   "dependencies": {
+    "@lumino/coreutils": "^1.5.3"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
