+++ packages/kernelspec/src/index.ts
@@ -2,3 +2,4 @@
 // Distributed under the terms of the Modified BSD License.
 
 export * from './kernelspecs';
+export * from './tokens';
