+++ packages/kernelspec/src/kernelspecs.ts
@@ -1,42 +1,72 @@
 import { KernelSpec } from '@jupyterlab/services';
 
+import { IKernelSpecs } from './tokens';
+
 /**
  * A class to handle requests to /api/kernelspecs
  */
-export class KernelSpecs {
+export class KernelSpecs implements IKernelSpecs {
+  /**
+   * Construct a new KernelSpecs.
+   *
+   * @param options The instantiation options.
+   */
+  constructor(options: KernelSpecs.IOptions) {
+    console.log(options);
+  }
+
   /**
    * Get the kernel specs.
    */
   get specs(): KernelSpec.ISpecModels | null {
-    return Private.DEFAULT_SPECS;
+    if (this._specs.size === 0) {
+      return null;
+    }
+    return {
+      default: 'javascript',
+      kernelspecs: Object.fromEntries(this._specs)
+    };
   }
+
+  /**
+   * Get the kernel factories for the current kernels.
+   */
+  get factories(): KernelSpecs.KernelFactories {
+    return this._factories;
+  }
+
+  /**
+   * Register a new kernel.
+   */
+  register(options: KernelSpecs.IKernelOptions): void {
+    const { spec, create } = options;
+    this._specs.set(spec.name, spec);
+    this._factories.set(spec.name, create);
+  }
+
+  private _specs = new Map<string, KernelSpec.ISpecModel>();
+  private _factories = new Map<string, KernelSpecs.KernelFactory>();
 }
 
 /**
- * A namespace for private data.
+ * A namespace for KernelSpecs statics.
  */
-namespace Private {
-  export const DEFAULT_SPECS: KernelSpec.ISpecModels = {
-    default: 'JavaScript',
-    kernelspecs: {
-      javascript: {
-        name: 'javascript',
-        display_name: 'JavaScript',
-        language: 'javascript',
-        argv: [],
-        spec: {
-          argv: [],
-          env: {},
-          display_name: 'JavaScript',
-          language: 'javascript',
-          interrupt_mode: 'message',
-          metadata: {}
-        },
-        resources: {
-          'logo-32x32': 'TODO',
-          'logo-64x64': '/kernelspecs/javascript.svg'
-        }
-      }
-    }
-  };
+export namespace KernelSpecs {
+  /**
+   * The instantiation options for a KernelSpecs
+   */
+  export interface IOptions {}
+
+  /**
+   * Registration options for a new kernel.
+   */
+  export interface IKernelOptions {
+    spec: KernelSpec.ISpecModel;
+    create: KernelFactory;
+  }
+
+  // TODO: define the options
+  export type KernelFactory = () => Promise<void>;
+
+  export type KernelFactories = Map<string, KernelFactory>;
 }
