+++ packages/kernelspec/src/tokens.ts
@@ -1,0 +1,34 @@
+import { KernelSpec } from '@jupyterlab/services';
+
+import { Token } from '@lumino/coreutils';
+
+import { KernelSpecs } from './kernelspecs';
+
+/**
+ * The token for the kernel spec service.
+ */
+export const IKernelSpecs = new Token<IKernelSpecs>(
+  '@jupyterlite/kernelspec:IKernelSpecs'
+);
+
+/**
+ * The interface for the kernel specs service.
+ */
+export interface IKernelSpecs {
+  /**
+   * Get the kernel specs.
+   */
+  readonly specs: KernelSpec.ISpecModels | null;
+
+  /**
+   * Get the kernel factories for the current kernels.
+   */
+  readonly factories: KernelSpecs.KernelFactories;
+
+  /**
+   * Register a new kernel spec
+   *
+   * @param options The kernel spec options.
+   */
+  register: (options: KernelSpecs.IKernelOptions) => void;
+}
