+++ tsconfigbase.json
@@ -16,7 +16,7 @@
     "strict": true,
     "skipLibCheck": true,
     "strictNullChecks": true,
-    "target": "es2017",
+    "target": "es2019",
     "types": []
   }
 }
